app/requirements.txt
mysql-connector-python
pulp
gunicorn